Put a try-catch-block inside hideAddItemTable() in categories.js and playlist.js
hideAddItemTable() is called by pagination.js every time the user clicks on one of the buttons on the paginator. The purpose of hideAddItemTable() is to ensure that, if the user clicked on the plus button and the row for adding an item is showing, the row for adding an item is replaced with an row with the button for adding an item. Also, I have replaced the icons in editItem() in categories.js with the same set of icons that is used in playlist.js